requirements.txt
aiogram==3.22.0
python-dotenv==1.0.0
aiosqlite==0.19.0

# --- Тестовые зависимости ---
pytest==8.2.0
pytest-asyncio==0.23.6
aiogram-tests==1.0.3